Issue #2665664 by amaria: Adds a flexslider responsive field formatter.
