+++ flexslider_fields/flexslider_fields.module
@@ -6,72 +6,15 @@
  *
  * @author jepedo
  * @author Mathew Winstone <mwinstone@coldfrontlabs.ca>
+ * @author Agnes Chisholm <amaria@66428.no-reply.drupal.org>
  */
 
 /**
- * Implements hook_field_formatter_view().
- *
- * Prepares a renderable array of images and adds the neccessary JS and CSS.
+ * Implements hook_field_formatter_info_alter().
  */
-function flexslider_fields_field_formatter_view($entity_type, $entity, $field, $instance, $langcode, $items, $display) {
-
-  $element = array();
-  if (count($items) > 0) {
-
-    foreach ($items as $key => $item) {
-      if (isset($item['item'])) {
-        continue;
-      }
-      $tmp = $item;
-      $item = array();
-      $item['item'] = $tmp;
-
-      // Setup the variables for calling theme_image_style.
-      if (isset($item['item']['uri'])) {
-        $item['slide']['path'] = $item['item']['uri'];
-      }
-      if (isset($display['settings']['image_style'])) {
-        $item['slide']['style_name'] = $display['settings']['image_style'];
-      }
-      if (isset($item['item']['width'])) {
-        $item['slide']['width'] = $item['item']['width'];
-      }
-      if (isset($item['item']['height'])) {
-        $item['slide']['height'] = $item['item']['height'];
-      }
-      if (isset($item['item']['alt'])) {
-        $item['slide']['alt'] = $item['item']['alt'];
-      }
-      if (isset($item['item']['title'])) {
-        $item['slide']['title'] = $item['item']['title'];
-      }
-
-      // Render the slide item
-      // If not style set, we have to call theme_image since theme_image_style
-      // doesn't auto-fallback to full size image.
-      if (!empty($item['slide']['style_name'])) {
-        // Generate the HTML for the slide.
-        $item['slide'] = theme('image_style', $item['slide']);
-      }
-      else {
-        // Generate the HTML for the slide.
-        $item['slide'] = theme('image', $item['slide']);
-      }
-
-      // Check caption settings.
-      if ($display['settings']['caption']) {
-        $item['caption'] = filter_xss($item['item']['title']);
-      }
-
-      $items[$key] = $item;
-    }
-
-    $element = array(
-      '#theme' => 'flexslider',
-      '#items' => $items,
-      '#settings' => $display['settings'],
-    );
+function flexslider_fields_field_formatter_info_alter(array &$info) {
+  // Remove the responsive formatter if responsive image module is not installed.
+  if (!\Drupal::moduleHandler()->moduleExists('responsive_image') && isset($info['flexslider_responsive'])) {
+    unset($info['flexslider_responsive']);
   }
-
-  return $element;
 }
